Added some customer application scenario descriptions. Safety commit before starting on order module.

+++ modules/inventory/inventory-core/src/main/java/org/ucommerce/modules/inventory/services/InventoryStock.java
@@ -3,5 +3,5 @@ package org.ucommerce.modules.inventory.services;
 import org.ucommerce.modules.inventory.model.LocationId;
 import org.ucommerce.shared.kernel.ids.ProductId;
 
-public record InventoryStock(LocationId locationId, ProductId productId, int stock) {
+public record InventoryStock(LocationId locationId, ProductId productId, long stock) {
 }
